auth-service/pyproject.toml
[project]
name = "smart-home-auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "donuwave",email = "d_tareev@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "aio-pika (>=9.5.5,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
